Rename gatsby plugin
