--- packages/demo-theme/gatsby-config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  __experimentalThemes: ['gatsby-theme-ui'],
-}
