--- packages/demo-theme/package.json
@@ -1,12 +1,0 @@
-{
-  "private": true,
-  "name": "demo-theme",
-  "version": "0.2.0",
-  "main": "gatsby-config.js",
-  "author": "Brent Jackson",
-  "license": "MIT",
-  "dependencies": {
-    "gatsby-theme-ui": "^0.2.0",
-    "theme-ui": "^0.2.0"
-  }
-}
