--- packages/demo-theme/src/gatsby-theme-ui/index.js
@@ -1,10 +1,0 @@
-export default {
-  initialColorMode: 'light',
-  colors: {
-    text: 'tomato',
-    background: 'papayawhip',
-    primary: 'rebeccapurple',
-    muted: 'peachpuff',
-    modes: {},
-  },
-}
