+++ packages/docs/gatsby-config.js
@@ -1,7 +1,7 @@
 const remarkPlugins = [require('remark-slug')]
 
 module.exports = {
-  __experimentalThemes: ['gatsby-theme-ui'],
+  __experimentalThemes: ['gatsby-plugin-theme-ui'],
   plugins: [
     'gatsby-plugin-react-helmet',
     'gatsby-plugin-catch-links',
