+++ packages/docs/package.json
@@ -28,7 +28,7 @@
     "gatsby-plugin-mdx": "^1.0.4",
     "gatsby-plugin-react-helmet": "^3.0.12",
     "gatsby-remark-prismjs": "^3.2.9",
-    "gatsby-theme-ui": "^0.2.0",
+    "gatsby-plugin-theme-ui": "^0.2.0",
     "lodash.merge": "^4.6.1",
     "lodash.omit": "^4.5.0",
     "prismjs": "^1.16.0",
