+++ packages/docs/src/pages/api.mdx
@@ -73,7 +73,7 @@ const context = useThemeUI()
 ### `ColorMode`
 
 The `ColorMode` component applies foreground and background color to the `<body>` element, based on the color modes defined in a theme.
-It is also used in `gatsby-theme-ui` for helping reduce flashes of style changes on page load.
+It is also used in `gatsby-plugin-theme-ui` for helping reduce flashes of style changes on page load.
 
 **Usage**
