+++ packages/docs/src/pages/gatsby-plugin.mdx
@@ -3,7 +3,7 @@ title: 'Gatsby Plugin'
 ---
 
 import Layout from '../components/layout'
-import Readme from 'gatsby-theme-ui/README.md'
+import Readme from 'gatsby-plugin-theme-ui/README.md'
 
 export default Layout
