+++ packages/docs/src/pages/notes.mdx
@@ -12,8 +12,5 @@ Other considerations for this UI library:
 
 - [ ] Icons
 - [ ] Handling webfonts
-- [ ] Color modes & `useColorMode` w/ `localStorage`
-- [ ] `useDarkMode` - hooks into media query
 - [ ] Base UI components
-- [ ] POC `gatsby-theme-ui`
 - [ ] `Sidebar` component + state
