+++ packages/gatsby-plugin-theme-ui/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "gatsby-theme-ui",
+  "name": "gatsby-plugin-theme-ui",
   "version": "0.2.0",
   "main": "index.js",
   "author": "Brent Jackson",
